🟢green(CrewGroup): 크루 그룹 객체 그린 테스트
